Move function into lib/executions.js Sort by timestamp for single granule
